requirements.txt
Flask==2.1.2
numpy==1.22.3
opencv_python==4.5.5.64
Pillow==9.1.0

Werkzeug~=2.1.2
